+++ celescope/atac/__init__.py
@@ -1,4 +1,4 @@
-STEPS = ["mkref", "sample", "barcode", "atac", "analysis"]
+STEPS = ["mkref", "sample", "barcode", "atac", "analysis", "cells"]
 __ASSAY__ = "atac"
 
 REMOVE_FROM_MULTI = {